+++ src/Infrastructure/Session/Http/ActiveSessionGuard.ts
@@ -1,13 +1,14 @@
 import { OK, type SAR } from '@hexancore/common';
 import { Injectable } from '@nestjs/common';
 import type { SessionData } from '../Data';
-import type { Session } from '../Session';
-import { ActiveSessionGuard } from './ActiveSessionGuard';
+import { Session } from '../Session';
+import { AbstractSessionGuard } from './AbstractSessionGuard';
 
 @Injectable()
-export class AuthenticatedSessionGuard extends ActiveSessionGuard {
+export class ActiveSessionGuard extends AbstractSessionGuard {
 
   protected canActivateWithSession<D extends SessionData = any>(session: Session<D>): SAR<boolean> {
-    return OK(session.isAuthenticated());
+    return OK(session.isActive());
   }
+
 }