+++ src/Infrastructure/Session/Http/types.ts
@@ -1,5 +1,5 @@
-import type { FRequest } from "@hexancore/core/.";
+import type { FRequest } from "@hexancore/core";
 import type { SessionData } from "../Data/SessionData";
 import type { Session } from "../Session";
 
-export type FReqWithSession<D extends SessionData> = FRequest & { session: Session<D> };
+export type FReqWithSession<D extends SessionData> = FRequest & { session?: Session<D>; };