+++ src/Test/Infrastructure/OpenId/App/Controller/TestOpenIdFrontendController.ts
@@ -1,9 +1,8 @@
 import { OK, type R } from "@hexancore/common";
 import { RedirectResult, type FRequest } from "@hexancore/core";
-import { Controller, Get, Header, Req } from "@nestjs/common";
+import { Controller, Get, Header, Req, UseInterceptors } from "@nestjs/common";
 import { ApiTags } from "@nestjs/swagger";
-import { type FReqWithSession, type OpenIdUserSessionData } from "../../../../../Infrastructure";
-
+import { SessionRequestInjectorInterceptor, type FReqWithSession, type OpenIdUserSessionData } from "../../../../../Infrastructure";
 
 @ApiTags('Test')
 @Controller({ path: 'front' })
@@ -14,20 +13,18 @@ export class TestOpenIdFrontendController {
   }
 
   @Get('protected/dashboard')
+  @UseInterceptors(SessionRequestInjectorInterceptor)
   @Header('Content-Type', 'text/html')
   public dashboard(@Req() req: FReqWithSession<OpenIdUserSessionData>): R<string | RedirectResult> {
-
     const session = req.session;
+    const cookiesHtml = req.cookies ? this.renderJsonPreHtml('Cookies', req.cookies) : '';
     if (!session) {
       return this.renderHtml({
         title: 'Dashboard',
         body: `
           <h3 style="text-align:center;">Dashboard</h3>
           <a class="button" href="${this.baseUrl}/user/public/auth/login">Login</a>
-          <p class="field">Cookies:</p>
-          <pre>
-          ${JSON.stringify(req.cookies, null, 2)}
-          </pre>
+          ${cookiesHtml}
         `
       });
     }
@@ -39,7 +36,6 @@ export class TestOpenIdFrontendController {
     <script>
     function loadCats() {
       const response = fetch("${this.baseUrl}/cat/protected/cats", {credentials: "include",}).then((response) => {
-        console.log(response);
         const container = document.getElementById('cats');
         container.innerHTML = '<h4>Cats</h4>';
 
@@ -62,6 +58,9 @@ export class TestOpenIdFrontendController {
 
     const loadCatsButton = `<button type="button" class="button" onClick="loadCats()">Load cats 😸</button>`;
 
+    const idTokenPayloadHtml = session.isAuthenticated() ? this.renderJsonPreHtml('IdTokenPayload', session.data!.auth!.idToken.getPayload()) : '';
+    const dataHtml = session.data ? this.renderJsonPreHtml('Data', session.data.toJSON()) : '';
+
     return this.renderHtml({
       title: 'Dashboard',
       body: `
@@ -75,25 +74,17 @@ export class TestOpenIdFrontendController {
           <h4>Session</h4>
           <p class="field">Id: <span>${session.id}</span></p>
           <p class="field">State: <span>${session.state}</span></p>
-          <p class="field">Created: <span id="sessionCreatedAt">${session.createdAt.formatDateTime()}Z</span></p>
-          <p class="field">Expire: <span id="sessionExpireAt">${session.expireAt.formatDateTime()}Z</span></p>
+          <p class="field">Created: <span id="sessionCreatedAt">${session.createdAt!.formatDateTime()}Z</span></p>
+          <p class="field">Expire: <span id="sessionExpireAt">${session.expireAt!.formatDateTime()}Z</span></p>
           <p class="field">isAuthenticated: <span>${session.isAuthenticated() ? 'Yes' : 'No'}</span></p>
           <p class="field">GroupId: <span>${session.getSessionGroupId()}</span></p>
-          <p class="field">Data:</p>
-          <pre>
-          ${JSON.stringify(session.data ? session.data.toJSON() : null, null, 2)}
-          </pre>
-
-          <p class="field">Cookies:</p>
-          <pre>
-          ${JSON.stringify(req.cookies, null, 2)}
-          </pre>
+          ${idTokenPayloadHtml}
+          ${dataHtml}
+          ${cookiesHtml}
         </div>
         <script>
         convertDateTimeToLocal('sessionCreatedAt');
         convertDateTimeToLocal('sessionExpireAt');
-
-
         </script>
       `
     });
@@ -102,19 +93,16 @@ export class TestOpenIdFrontendController {
   @Get('public/auth-error')
   @Header('Content-Type', 'text/html')
   public authError(@Req() req: FRequest): R<string> {
+    const queryHtml = req.query ? this.renderJsonPreHtml('Query', req.query) : '';
+    const cookiesHtml = req.cookies ? this.renderJsonPreHtml('Cookies', req.cookies) : '';
+
     return this.renderHtml({
       title: 'Authorize Error',
       body: `
         <h3 style="text-align:center;">Authorize error</h3>
         <a class="button" href="${this.baseUrl}/user/public/auth/login">Retry Login</a>
-        <p class="field">Query:</p>
-        <pre>
-        ${JSON.stringify(req.query, null, 2)}
-        </pre>
-        <p class="field">Cookies:</p>
-        <pre>
-        ${JSON.stringify(req.cookies, null, 2)}
-        </pre>
+        ${queryHtml}
+        ${cookiesHtml}
       `
     });
   }
@@ -122,17 +110,23 @@ export class TestOpenIdFrontendController {
   @Get('public/post-logout')
   @Header('Content-Type', 'text/html')
   public postLogout(@Req() req: FRequest): R<string> {
+    const cookiesHtml = req.cookies ? this.renderJsonPreHtml('Cookies', req.cookies) : '';
     return this.renderHtml({
       title: 'Post Logout',
       body: `
         <h3 style="text-align:center;">Post Logout</h3>
         <a class="button" href="${this.baseUrl}/user/public/auth/login">Login</a>
-        <p class="field">Cookies:</p>
-        <pre>
-          ${JSON.stringify(req.cookies, null, 2)}
-        </pre>
+        ${cookiesHtml}
       `
     });
+  }
+
+  private renderJsonPreHtml(header: string, data: any): string {
+    const text = JSON.stringify(data, null, 2);
+    return `
+    <p class="field">${header}:</p>
+    <pre>${text}</pre>
+    `;
   }
 
   private renderHtml(context: Record<string, any>) {
@@ -150,6 +144,11 @@ export class TestOpenIdFrontendController {
             margin: auto;
             width: 80%;
           }
+          pre {
+            border: solid 1px black;
+            padding: 16px;
+          }
+
           p.field {
             font-weight: bold;
           }
@@ -170,13 +169,13 @@ export class TestOpenIdFrontendController {
             border-radius: 7px;
             display: inline-block;
            }
-           .button:hover{
+          .button:hover{
             background: linear-gradient(to right,#5482d0 ,#7d5ee3);
             background-color: #5482d0;
-           }
-           .button:active{
+          }
+          .button:active{
             transform: scale(0.95);
-           }
+          }
 
         </style>
         ${context.head ?? ''}