+++ src/Test/Infrastructure/OpenId/index.ts
@@ -1,4 +1,2 @@
 export * from './Provider/TestOpenIdProviderServer';
 export * from './Provider/TestOpenIdConfigurationBuilder';
-
-export * from './App/TestOpenIdAppModule';