+++ CHANGELOG.md
@@ -7,13 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-06-20
+
+### Added
+
+- added more exports options in package.json
+
+### Changed
+
+- upgraded dependencies
+- reworked session guards
+- reworked SessionInterceptor
+
 ## [0.3.0] - 2024-05-16
 
 ### Changed
 
 - Reworked OpenId.
 - Reworked Session.
-- Upgrade dependencies.
+- Upgrade dependencies
 
 ## [0.2.0] - 2024-04-24
 
@@ -34,8 +46,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Inital release
 
-[unreleased] https://github.com/hexancore/auth/compare/0.3.0...HEAD   
-[0.3.0] https://github.com/hexancore/auth/compare/0.2.0...0.3.0   
-[0.2.0] https://github.com/hexancore/auth/compare/0.1.1...0.2.0   
+[unreleased] https://github.com/hexancore/auth/compare/0.4.0...HEAD   
+[0.4.0] https://github.com/hexancore/auth/compare/0.3.0...0.4.0   
+[0.3.0] https://github.com/hexancore/auth/compare/0.2.0...0.3.0  
+[0.2.0] https://github.com/hexancore/auth/compare/0.1.1...0.2.0  
 [0.1.1] https://github.com/hexancore/auth/compare/0.1.0...0.1.1  
 [0.1.0] https://github.com/hexancore/auth/releases/tag/0.1.0